moved message logic from Backend to ErrBot.
