+++ errbot/backends/base.py
@@ -1,10 +1,8 @@
-import difflib
 import inspect
 import io
 import logging
 import random
 import time
-import traceback
 import warnings
 
 from collections import deque, defaultdict
@@ -12,9 +10,7 @@ from xml.etree import cElementTree as ET
 from xml.etree.cElementTree import ParseError
 
 from errbot import botcmd, PY2
-from errbot.utils import get_sender_username, xhtml2txt, split_string_after, deprecated
-from errbot.templating import tenv
-from errbot.bundled.threadpool import ThreadPool, WorkRequest
+from errbot.utils import get_sender_username, xhtml2txt, deprecated
 
 log = logging.getLogger(__name__)
 
@@ -595,7 +591,7 @@ class Backend(object):
     MSG_ERROR_OCCURRED = 'Sorry for your inconvenience. ' \
                          'An unexpected error occurred.'
 
-    def __init__(self, config):
+    def __init__(self, _):
         """ Those arguments will be directly those put in BOT_IDENTITY
         """
         log.debug("Backend init.")
@@ -606,25 +602,8 @@ class Backend(object):
         self._reconnection_multiplier = 1.75  # Delay multiplier
         self._reconnection_jitter = (0, 3)    # Random jitter added to delay (min, max)
 
-        if config.BOT_ASYNC:
-            self.thread_pool = ThreadPool(3)
-            log.debug('created the thread pool' + str(self.thread_pool))
-        self.commands = {}  # the dynamically populated list of commands available on the bot
-        self.re_commands = {}  # the dynamically populated list of regex-based commands available on the bot
-        self.MSG_UNKNOWN_COMMAND = 'Unknown command: "%(command)s". ' \
-                                   'Type "' + config.BOT_PREFIX + 'help" for available commands.'
-        if config.BOT_ALT_PREFIX_CASEINSENSITIVE:
-            self.bot_alt_prefixes = tuple(prefix.lower() for prefix in config.BOT_ALT_PREFIXES)
-        else:
-            self.bot_alt_prefixes = config.BOT_ALT_PREFIXES
-
     def send_message(self, mess):
-        """Should be overridden by backends"""
-        raise NotImplementedError("send_message should be implemented by the backend")
-
-    def send_simple_reply(self, mess, text, private=False):
-        """Send a simple response to a message"""
-        self.send_message(self.build_reply(mess, text, private))
+        """Should be overridden by backends with a super().send_message() call."""
 
     def build_reply(self, mess, text=None, private=False):
         """ Should be implemented by the backend """
@@ -653,406 +632,6 @@ class Backend(object):
             See :class:`~errbot.errBot.ErrBot`
         """
         pass
-
-    def callback_message(self, mess):
-        """
-        Needs to return False if we want to stop further treatment
-        """
-        # Prepare to handle either private chats or group chats
-        type_ = mess.type
-        frm = mess.frm
-        text = mess.body
-        if not hasattr(mess.frm, 'person'):
-            raise Exception('mess.frm not an Identifier as it misses the "person" property. Class of frm : %s'
-                            % mess.frm.__class__)
-
-        username = mess.frm.person
-        user_cmd_history = self.cmd_history[username]
-
-        if mess.delayed:
-            log.debug("Message from history, ignore it")
-            return False
-
-        if type_ not in ("groupchat", "chat"):
-            log.debug("unhandled message type %s" % mess)
-            return False
-
-        if (frm.person == self.bot_identifier.person or
-            type_ == "groupchat" and mess.nick == self.bot_config.CHATROOM_FN):  # noqa
-                log.debug("Ignoring message from self")
-                return False
-
-        log.debug("*** frm = %s" % frm)
-        log.debug("*** username = %s" % username)
-        log.debug("*** type = %s" % type_)
-        log.debug("*** text = %s" % text)
-
-        # If a message format is not supported (eg. encrypted),
-        # txt will be None
-        if not text:
-            return False
-
-        surpress_cmd_not_found = False
-
-        prefixed = False  # Keeps track whether text was prefixed with a bot prefix
-        only_check_re_command = False  # Becomes true if text is determed to not be a regular command
-        tomatch = text.lower() if self.bot_config.BOT_ALT_PREFIX_CASEINSENSITIVE else text
-        if len(self.bot_config.BOT_ALT_PREFIXES) > 0 and tomatch.startswith(self.bot_alt_prefixes):
-            # Yay! We were called by one of our alternate prefixes. Now we just have to find out
-            # which one... (And find the longest matching, in case you have 'err' and 'errbot' and
-            # someone uses 'errbot', which also matches 'err' but would leave 'bot' to be taken as
-            # part of the called command in that case)
-            prefixed = True
-            longest = 0
-            for prefix in self.bot_alt_prefixes:
-                l = len(prefix)
-                if tomatch.startswith(prefix) and l > longest:
-                    longest = l
-            log.debug("Called with alternate prefix '{}'".format(text[:longest]))
-            text = text[longest:]
-
-            # Now also remove the separator from the text
-            for sep in self.bot_config.BOT_ALT_PREFIX_SEPARATORS:
-                # While unlikely, one may have separators consisting of
-                # more than one character
-                l = len(sep)
-                if text[:l] == sep:
-                    text = text[l:]
-        elif type_ == "chat" and self.bot_config.BOT_PREFIX_OPTIONAL_ON_CHAT:
-            log.debug("Assuming '%s' to be a command because BOT_PREFIX_OPTIONAL_ON_CHAT is True" % text)
-            # In order to keep noise down we surpress messages about the command
-            # not being found, because it's possible a plugin will trigger on what
-            # was said with trigger_message.
-            surpress_cmd_not_found = True
-        elif not text.startswith(self.bot_config.BOT_PREFIX):
-            only_check_re_command = True
-        if text.startswith(self.bot_config.BOT_PREFIX):
-            text = text[len(self.bot_config.BOT_PREFIX):]
-            prefixed = True
-
-        text = text.strip()
-        text_split = text.split(' ')
-        cmd = None
-        command = None
-        args = ''
-        if not only_check_re_command:
-            if len(text_split) > 1:
-                command = (text_split[0] + '_' + text_split[1]).lower()
-                if command in self.commands:
-                    cmd = command
-                    args = ' '.join(text_split[2:])
-
-            if not cmd:
-                command = text_split[0].lower()
-                args = ' '.join(text_split[1:])
-                if command in self.commands:
-                    cmd = command
-                    if len(text_split) > 1:
-                        args = ' '.join(text_split[1:])
-
-            if command == self.bot_config.BOT_PREFIX:  # we did "!!" so recall the last command
-                if len(user_cmd_history):
-                    cmd, args = user_cmd_history[-1]
-                else:
-                    return False  # no command in history
-            elif command.isdigit():  # we did "!#" so we recall the specified command
-                index = int(command)
-                if len(user_cmd_history) >= index:
-                    cmd, args = user_cmd_history[-index]
-                else:
-                    return False  # no command in history
-
-        # Try to match one of the regex commands if the regular commands produced no match
-        matched_on_re_command = False
-        if not cmd:
-            if prefixed:
-                commands = self.re_commands
-            else:
-                commands = {k: self.re_commands[k] for k in self.re_commands
-                            if not self.re_commands[k]._err_command_prefix_required}
-
-            for name, func in commands.items():
-                if func._err_command_matchall:
-                    match = list(func._err_command_re_pattern.finditer(text))
-                else:
-                    match = func._err_command_re_pattern.search(text)
-                if match:
-                    log.debug("Matching '{}' against '{}' produced a match"
-                              .format(text, func._err_command_re_pattern.pattern))
-                    matched_on_re_command = True
-                    self._process_command(mess, name, text, match)
-                else:
-                    log.debug("Matching '{}' against '{}' produced no match"
-                              .format(text, func._err_command_re_pattern.pattern))
-        if matched_on_re_command:
-            return True
-
-        if cmd:
-            self._process_command(mess, cmd, args, match=None)
-        elif not only_check_re_command:
-            log.debug("Command not found")
-            if surpress_cmd_not_found:
-                log.debug("Surpressing command not found feedback")
-            else:
-                reply = self.unknown_command(mess, command, args)
-                if reply is None:
-                    reply = self.MSG_UNKNOWN_COMMAND % {'command': command}
-                if reply:
-                    self.send_simple_reply(mess, reply)
-        return True
-
-    def _process_command(self, mess, cmd, args, match):
-        """Process and execute a bot command"""
-
-        frm = mess.frm
-        username = frm.person
-        user_cmd_history = self.cmd_history[username]
-
-        log.info("Processing command '{}' with parameters '{}' from {}".format(cmd, args, frm))
-
-        if (cmd, args) in user_cmd_history:
-            user_cmd_history.remove((cmd, args))  # Avoids duplicate history items
-
-        try:
-            self.check_command_access(mess, cmd)
-        except ACLViolation as e:
-            if not self.bot_config.HIDE_RESTRICTED_ACCESS:
-                self.send_simple_reply(mess, str(e))
-            return
-
-        f = self.re_commands[cmd] if match else self.commands[cmd]
-
-        if f._err_command_admin_only and self.bot_config.BOT_ASYNC:
-            # If it is an admin command, wait until the queue is completely depleted so
-            # we don't have strange concurrency issues on load/unload/updates etc...
-            self.thread_pool.wait()
-
-        if f._err_command_historize:
-            user_cmd_history.append((cmd, args))  # add it to the history only if it is authorized to be so
-
-        # Don't check for None here as None can be a valid argument to str.split.
-        # '' was chosen as default argument because this isn't a valid argument to str.split()
-        if not match and f._err_command_split_args_with != '':
-            try:
-                if hasattr(f._err_command_split_args_with, "parse_args"):
-                    args = f._err_command_split_args_with.parse_args(args)
-                elif callable(f._err_command_split_args_with):
-                    args = f._err_command_split_args_with(args)
-                else:
-                    args = args.split(f._err_command_split_args_with)
-            except Exception as e:
-                self.send_simple_reply(
-                    mess,
-                    "Sorry, I couldn't parse your arguments. {}".format(e)
-                )
-                return
-
-        if self.bot_config.BOT_ASYNC:
-            wr = WorkRequest(
-                self._execute_and_send,
-                [],
-                {'cmd': cmd, 'args': args, 'match': match, 'mess': mess,
-                 'template_name': f._err_command_template}
-            )
-            self.thread_pool.putRequest(wr)
-            if f._err_command_admin_only:
-                # Again, if it is an admin command, wait until the queue is completely
-                # depleted so we don't have strange concurrency issues.
-                self.thread_pool.wait()
-        else:
-            self._execute_and_send(cmd=cmd, args=args, match=match, mess=mess,
-                                   template_name=f._err_command_template)
-
-    def _execute_and_send(self, cmd, args, match, mess, template_name=None):
-        """Execute a bot command and send output back to the caller
-
-        cmd: The command that was given to the bot (after being expanded)
-        args: Arguments given along with cmd
-        match: A re.MatchObject if command is coming from a regex-based command, else None
-        mess: The message object
-        template_name: The name of the template which should be used to render
-            html-im output, if any
-
-        """
-        def process_reply(reply_):
-            # integrated templating
-            if template_name:
-                reply_ = tenv().get_template(template_name + '.html').render(**reply_)
-
-            # Reply should be all text at this point (See https://github.com/gbin/err/issues/96)
-            return str(reply_)
-
-        def send_reply(reply_):
-            for part in split_string_after(reply_, self.bot_config.MESSAGE_SIZE_LIMIT):
-                self.send_simple_reply(mess, part, cmd in self.bot_config.DIVERT_TO_PRIVATE)
-
-        commands = self.re_commands if match else self.commands
-        try:
-            if inspect.isgeneratorfunction(commands[cmd]):
-                replies = commands[cmd](mess, match) if match else commands[cmd](mess, args)
-                for reply in replies:
-                    if reply:
-                        send_reply(process_reply(reply))
-            else:
-                reply = commands[cmd](mess, match) if match else commands[cmd](mess, args)
-                if reply:
-                    send_reply(process_reply(reply))
-        except Exception as e:
-            tb = traceback.format_exc()
-            log.exception('An error happened while processing '
-                          'a message ("%s"): %s"' %
-                          (mess.body, tb))
-            send_reply(self.MSG_ERROR_OCCURRED + ':\n %s' % e)
-
-    def is_admin(self, usr):
-        """
-        an overridable check to see if a user is an administrator
-        """
-        return usr in self.bot_config.BOT_ADMINS
-
-    def check_command_access(self, mess, cmd):
-        """
-        Check command against ACL rules
-
-        Raises ACLViolation() if the command may not be executed in the given context
-        """
-        usr = str(mess.frm.person)
-        typ = mess.type
-
-        if cmd not in self.bot_config.ACCESS_CONTROLS:
-            self.bot_config.ACCESS_CONTROLS[cmd] = self.bot_config.ACCESS_CONTROLS_DEFAULT
-
-        if ('allowusers' in self.bot_config.ACCESS_CONTROLS[cmd] and
-           usr not in self.bot_config.ACCESS_CONTROLS[cmd]['allowusers']):
-            raise ACLViolation("You're not allowed to access this command from this user")
-        if ('denyusers' in self.bot_config.ACCESS_CONTROLS[cmd] and
-           usr in self.bot_config.ACCESS_CONTROLS[cmd]['denyusers']):
-            raise ACLViolation("You're not allowed to access this command from this user")
-        if typ == 'groupchat':
-            if not hasattr(mess.frm, 'room'):
-                raise Exception('mess.frm is not a MUCIdentifier as it misses the "room" property. Class of frm : %s'
-                                % mess.frm.__class__)
-            room = str(mess.frm.room)
-            if ('allowmuc' in self.bot_config.ACCESS_CONTROLS[cmd] and
-               self.bot_config.ACCESS_CONTROLS[cmd]['allowmuc'] is False):
-                raise ACLViolation("You're not allowed to access this command from a chatroom")
-            if ('allowrooms' in self.bot_config.ACCESS_CONTROLS[cmd] and
-               room not in self.bot_config.ACCESS_CONTROLS[cmd]['allowrooms']):
-                raise ACLViolation("You're not allowed to access this command from this room")
-            if ('denyrooms' in self.bot_config.ACCESS_CONTROLS[cmd] and
-               room in self.bot_config.ACCESS_CONTROLS[cmd]['denyrooms']):
-                raise ACLViolation("You're not allowed to access this command from this room")
-        else:
-            if ('allowprivate' in self.bot_config.ACCESS_CONTROLS[cmd] and
-               self.bot_config.ACCESS_CONTROLS[cmd]['allowprivate'] is False):
-                raise ACLViolation("You're not allowed to access this command via private message to me")
-
-        f = self.commands[cmd] if cmd in self.commands else self.re_commands[cmd]
-
-        if f._err_command_admin_only:
-            if typ == 'groupchat':
-                raise ACLViolation("You cannot administer the bot from a chatroom, message the bot directly")
-            if not self.is_admin(usr):
-                raise ACLViolation("This command requires bot-admin privileges")
-
-    def unknown_command(self, _, cmd, args):
-        """ Override the default unknown command behavior
-        """
-        full_cmd = cmd + ' ' + args.split(' ')[0] if args else None
-        if full_cmd:
-            part1 = 'Command "%s" / "%s" not found.' % (cmd, full_cmd)
-        else:
-            part1 = 'Command "%s" not found.' % cmd
-        ununderscore_keys = [m.replace('_', ' ') for m in self.commands.keys()]
-        matches = difflib.get_close_matches(cmd, ununderscore_keys)
-        if full_cmd:
-            matches.extend(difflib.get_close_matches(full_cmd, ununderscore_keys))
-        matches = set(matches)
-        if matches:
-            return (part1 + '\n\nDid you mean "' + self.bot_config.BOT_PREFIX +
-                    ('" or "' + self.bot_config.BOT_PREFIX).join(matches) + '" ?')
-        else:
-            return part1
-
-    def inject_commands_from(self, instance_to_inject):
-        classname = instance_to_inject.__class__.__name__
-        for name, value in inspect.getmembers(instance_to_inject, inspect.ismethod):
-            if getattr(value, '_err_command', False):
-                commands = self.re_commands if getattr(value, '_err_re_command') else self.commands
-                name = getattr(value, '_err_command_name')
-
-                if name in commands:
-                    f = commands[name]
-                    new_name = (classname + '-' + name).lower()
-                    self.warn_admins('%s.%s clashes with %s.%s so it has been renamed %s' % (
-                        classname, name, type(f.__self__).__name__, f.__name__, new_name))
-                    name = new_name
-                commands[name] = value
-
-                if getattr(value, '_err_re_command'):
-                    log.debug('Adding regex command : %s -> %s' % (name, value.__name__))
-                    self.re_commands = commands
-                else:
-                    log.debug('Adding command : %s -> %s' % (name, value.__name__))
-                    self.commands = commands
-
-    def remove_commands_from(self, instance_to_inject):
-        for name, value in inspect.getmembers(instance_to_inject, inspect.ismethod):
-            if getattr(value, '_err_command', False):
-                name = getattr(value, '_err_command_name')
-                if getattr(value, '_err_re_command') and name in self.re_commands:
-                    del (self.re_commands[name])
-                elif not getattr(value, '_err_re_command') and name in self.commands:
-                    del (self.commands[name])
-
-    def warn_admins(self, warning):
-        for admin in self.bot_config.BOT_ADMINS:
-            self.send(admin, warning)
-
-    def top_of_help_message(self):
-        """Returns a string that forms the top of the help message
-
-        Override this method in derived class if you
-        want to add additional help text at the
-        beginning of the help message.
-        """
-        return ""
-
-    def bottom_of_help_message(self):
-        """Returns a string that forms the bottom of the help message
-
-        Override this method in derived class if you
-        want to add additional help text at the end
-        of the help message.
-        """
-        return ""
-
-    def send(self, user, text, in_reply_to=None, message_type='chat', groupchat_nick_reply=False):
-        """Sends a simple message to the specified user.
-        """
-        if isinstance(user, str):
-            log.warn("""deprecation notice: you passed user as a string.
-                    You should use self.build_identifier(user) instead or directly an identifier if you have one.""")
-            user = self.build_identifier(user)
-
-        nick_reply = self.bot_config.GROUPCHAT_NICK_PREFIXED
-
-        if (message_type == 'groupchat' and in_reply_to and nick_reply and groupchat_nick_reply):
-            reply_text = self.groupchat_reply_format().format(in_reply_to.nick, text)
-        else:
-            reply_text = text
-
-        mess = self.build_message(reply_text)
-        mess.to = user
-
-        if in_reply_to:
-            mess.type = in_reply_to.type
-            mess.frm = in_reply_to.to
-        else:
-            mess.type = message_type
-            mess.frm = self.bot_identifier
-
-        self.send_message(mess)
 
     def serve_forever(self):
         """
